Change: Made several changes and additions relating to keybinds
The priamry focus of this was to add playerctl bindings and make better,
easier to use, left hand only bindings for common keybinds, like swaping windows
between screens.

Addition: Create playerctl bindings for play-pause, next, previous track
Change: Help function is now 'C-/', which uses the question mark ? to mind you
where it is.

Note: Here is a warning about the nature of awesome, many libraries are included
in the global table, and will easily cause confusion, and are not easily
debuggable either.
Top level tables like 'screen' are in the global table, overriding things will
mess them up.
These top level tables are accessible in most High Order Functions, but the
global table may not be.
Equally, you cannot easily find where the function will be called, so assume
it does not have access to the surrounding scope.